refactor: restructure pad folder
