+++ src/frontend/src/pad/Dashboard.scss
