+++ src/frontend/src/pad/StateIndicator.scss
