+++ src/frontend/src/pad/Terminal.scss
