+++ src/frontend/src/pad/Terminal.tsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect, useCallback, useRef } from 'react';
-import { useWorkspaceState } from '../../api/hooks';
+import { useWorkspaceState } from '../api/hooks';
 import type { NonDeleted, ExcalidrawEmbeddableElement } from '@atyrode/excalidraw/element/types';
 import type { AppState } from '@atyrode/excalidraw/types';
 import './Terminal.scss';
