+++ src/frontend/src/pad/buttons/ControlButton.scss
