+++ src/frontend/src/pad/index.ts
@@ -1,13 +1,13 @@
 // Re-export all components from the pad module
-export * from './controls/ControlButton';
-export * from './controls/StateIndicator';
-export * from './containers/Dashboard';
-export * from './containers/Terminal';
+export * from './buttons/ControlButton';
+export * from './StateIndicator';
+export * from './Dashboard';
+export * from './Terminal';
 export * from './buttons';
 export * from './editors';
 
 // Default exports
-export { default as ControlButton } from './controls/ControlButton';
-export { default as StateIndicator } from './controls/StateIndicator';
-export { default as Dashboard } from './containers/Dashboard';
-export { default as Terminal } from './containers/Terminal';
+export { default as ControlButton } from './buttons/ControlButton';
+export { default as StateIndicator } from './StateIndicator';
+export { default as Dashboard } from './Dashboard';
+export { default as Terminal } from './Terminal';
